requirements.txt
python_version >= "3.9" and python_version < "4.0"
anytree==2.12.1 ; python_version >= "3.9" and python_version < "4"
awkward-cpp==34 ; python_version >= "3.9" and python_version < "4.0"
awkward==2.6.5 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
cramjam==2.8.3 ; python_version >= "3.9" and python_version < "4.0"
cupy-cuda11x==13.1.0 ; python_version >= "3.9" and python_version < "4.0"
cython==3.0.10 ; python_version >= "3.9" and python_version < "4.0"
fastparquet==2024.5.0 ; python_version >= "3.9" and python_version < "4.0"
fastrlock==0.8.2 ; python_version >= "3.9" and python_version < "4.0"
fsspec==2024.6.0 ; python_version >= "3.9" and python_version < "4.0"
importlib-metadata==7.1.0 ; python_version >= "3.9" and python_version < "3.12"
joblib==1.4.2 ; python_version >= "3.9" and python_version < "4.0"
lark==1.1.9 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.9" and python_version < "4.0"
orjson==3.10.4 ; python_version >= "3.9" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.9" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.9" and python_version < "4.0"
pathlib==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
pyarrow==16.1.0 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.9" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
ruamel-yaml-clib==0.2.8 ; platform_python_implementation == "CPython" and python_version < "3.13" and python_version >= "3.9"
ruamel-yaml==0.18.6 ; python_version >= "3.9" and python_version < "4.0"
scipy==1.13.1 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4"
tqdm==4.66.4 ; python_version >= "3.9" and python_version < "4.0"
tree-maker @ git+https://github.com/xsuite/tree_maker.git@5eced09a6b57dc70a3eadd99197bd9c7afeeaec7 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.9" and python_version < "3.11"
tzdata==2024.1 ; python_version >= "3.9" and python_version < "4.0"
xcoll==0.4.0 ; python_version >= "3.9" and python_version < "4.0"
xdeps==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
xfields==0.17.0 ; python_version >= "3.9" and python_version < "4.0"
xobjects==0.4.1 ; python_version >= "3.9" and python_version < "4.0"
xpart==0.18.4 ; python_version >= "3.9" and python_version < "4.0"
xsuite==0.12.1 ; python_version >= "3.9" and python_version < "4.0"
xtrack==0.61.0 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.19.2 ; python_version >= "3.9" and python_version < "3.12"
